refactor: extract Game Instance Recovery Service from GameManager
Extract game recovery and database restoration operations into
GameInstanceRecoveryService to complete the major service extractions
from GameManager following the established refactoring plan.

Changes:
- Create GameInstanceRecoveryService with extracted recovery operations
  - recoverGameInstance() with full database deserialization
  - restoreMapDataToManager() and deserializeMapTiles()
  - loadGame() for testing integration
- Update GameManager to delegate to GameInstanceRecoveryService
- Register service in ServiceRegistry for proper lifecycle management
- Keep convertGeneratorType() in GameManager as it's still used for map generation
- Maintain all existing functionality for game recovery and persistence

Progress:
- GameManager reduced from 1,500 to 1,272 lines (228 line reduction)
- Total progress: 792 lines extracted from original 2,064 lines (38% reduction)
- Successfully completed major service extractions phase of refactoring
